Autoconf/make change and source refactoring for separate perf tests.
Move the timed performance tests out of the unit test driver
(pveclib_test).  Create a standalone executable (pveclib_perf) that
can be run for timed measurements and instruction trace analysts.
Trick is to avoid unintended function in-lining or loop unrolling
that might result in missing symbols (instruction trace targets)
optimizing to dead code. The plan goes: Functions that we want to
time are expanded into the corresponding vec_<type>_dummy.c and
called via extern. Timed loops calling the timed functions are
collected in the corresponding vec_pref_<type>.c files. The timed
loops are called from pveclib_perf.c. Each call is braceted by time
stamps. The Delta time (in time-base ticks and micro-seconds) is
reported. Reports are grouped by element type then operation.
Also updated configure.ac for the pending move to autoconf-1.71.

	* Makefile.in: Regenerated.
	* aclocal.m4: Regenerated.
	* configure: Regenerated.
	* src/Makefile.in: Regenerated.

	* configure.ac [AC_PREREQ]: Change to 2.69 prereq.
	[LT_INIT]: Replace AC_PROG_LIBTOOL.

	* src/Makefile.am [pveclib_test_SOURCES]: Remove
	testsuite/vec_perf_f64.c, testsuite/vec_perf_f32.c,
	testsuite/vec_perf_i512.c, testsuite/vec_perf_i128.c,
	testsuite/vec_perf_f32.h, testsuite/vec_perf_f64.h,
	testsuite/vec_perf_i512.h, testsuite/vec_perf_i128.h,
	[TESTS]: Add pveclib_perf.
	[pveclib_perf_SOURCES]: New sources starting with
	testsuite/pveclib_perf.c and others copied from
	pveclib_test_SOURCES.
	[pveclib_perf_CFLAGS]: New pveclib_perf CFLAGS.
	[pveclib_perf_LDADD]: New pveclib_perf LD adds.

	* src/testsuite/arith128_test_f128.c (test_scalar_add128,
	test_scalar_div128, test_scalar_mul128, test_scalar_sub128,
	test_scalarCC_expxsuba_128): Remove timed test externs.
	Moved to vec_perf_f128.c.
	(f128_one, f128_e, inv_fact2, inv_fact3, inv_fact4, inv_fact5,
	inv_fact6, inv_fact7, inv_fact8): Remove QP Float consts.
	Moved to vec_perf_f128.c.
	(test_scalarLib_expxsuba_128): Remove timed test function.
	Moved to vec_perf_f128.c.
	(test_vec_max8_f128, test_vec_max8_f128uz, test_gcc_max8_f128,
	test_vec_cmpgtuqp): Remove timed test externs.
	(test_lib_max8_f128): Remove timed test function.
	(timed_lib_max8_f128, timed_gcc_max8_f128, timed_vec_max8_f128,
	timed_vec_max8_f128uz, timed_expxsuba_v1_f128,
	timed_expxsuba_v2_f128): Remove timed test loop function.
	Moved to vec_perf_f128.c.
	(test_time_f128): Remove timed test driver. Moved to
	pveclib_perf.c.
	(test_extract_insert_f128): Remove call to test_vec_max8_f128.
	(test_vec_f128): Remove call to test_time_f128.

	* src/testsuite/arith128_test_f32.c:
	Remove #include <testsuite/vec_perf_f32.h>.
	(matrix_f32): remove static array.
	(test_f32_Imatrix_init, test_f32_matrix_transpose,
	test_f32_matrix_gather_transpose,
	test_f32_matrix_gatherx2_transpose,
	test_f32_matrix_gatherx4_transpose):
	Remove timed test functions. Moved to vec_f32_dummy.c.
	(test_time_f32): Remove timed test driver. Moved to
	pveclib_perf.c.
	(test_vec_f32): Remove call to test_time_f32.

	* src/testsuite/arith128_test_f64.c:
	Remove #include <testsuite/vec_perf_f64.h>.
	(matrix_f64): remove static array.
	(test_f64_Imatrix_init, test_f64_matrix_transpose,
	test_f64_matrix_gather_transpose,
	test_f64_matrix_gatherx2_transpose,
	test_f64_matrix_gatherx4_transpose):
	Remove timed test functions.
	(test_time_f64): Remove timed test driver. Moved to
	pveclib_perf.c.
	(test_vec_f64): Remove call to test_time_f64.

	* src/testsuite/arith128_test_i128.c:
	Remove #include <testsuite/vec_perf_i128.h>.
	(db_vec_clzq): Demove dead code.
	(test_time_i128): Remove timed test driver. Moved to
	pveclib_perf.c.
	(test_vec_i128): Remove call to test_time_i128.
	(example_longdiv_10e32): Remove extern to dummy code.
	* src/testsuite/arith128_test_i128.h
	(__test_muludq, example_longdiv_10e32, example_longbcdcf_10e32,
	example_longbcdct_10e32): Add extern to dummy code.

	* src/testsuite/arith128_test_i512.c:
	Remove #include <testsuite/vec_perf_i512.h>.
	(vec512_zeros, vec512_one, vec512_foxes, vec512_foxeasy,
	vec512_ten128th, vec512_ten256_h, vec512_ten256_l,
	vec512_ten384_h, vec512_ten384_m, vec512_ten384_l,
	vec512_ten512_3, vec512_ten512_2, vec512_ten512_1,
	vec512_ten512_0, vec512_ten1024_6, vec512_ten1024_5,
	vec512_ten1024_4, vec512_ten1024_3, vec512_ten1024_2,
	vec512_ten1024_1, vec512_ten1024_0, vec512_ten2048_13,
	vec512_ten2048_12, vec512_ten2048_11, vec512_ten2048_10,
	vec512_ten2048_9, vec512_ten2048_8, vec512_ten2048_7,
	vec512_ten2048_6, vec512_ten2048_5, vec512_ten2048_4,
	vec512_ten2048_3, vec512_ten2048_2, vec512_ten2048_1,
	vec512_ten2048_0): Remove multiple quadword integer consts.
	Moved to vec_int512_dummy.c.
	(test_time_i512): Remove timed test driver. Moved to
	pveclib_perf.c.
	(test_mul128x128): Remove call to test_time_i512.

	* src/testsuite/pveclib_perf.c: New file.

	* src/testsuite/vec_f128_dummy.c
	(test_convert_dpqp): New function.
	Multiple workarounds for clang.

	* src/testsuite/vec_f32_dummy.c
	Add #include <testsuite/arith128_test_f32.h>.
	(matrix_f32): Add static array.
	(test_f32_Imatrix_init, test_f32_matrix_transpose,
	test_f32_matrix_gather_transpose,
	test_f32_matrix_gatherx2_transpose,
	test_f32_matrix_gatherx4_transpose):
	Add timed test functions.

	* src/testsuite/vec_f64_dummy.c
	Add #include <testsuite/arith128_test_f64.h>.
	(matrix_f64): Add static array.
	(test_f64_Imatrix_init, test_f64_matrix_transpose,
	test_f64_matrix_gather_transpose,
	test_f64_matrix_gatherx2_transpose,
	test_f64_matrix_gatherx4_transpose):
	Add timed test functions.

	* src/testsuite/vec_int512_dummy.c
	(vec512_zeros, vec512_one, vec512_foxes, vec512_foxeasy,
	vec512_ten128th, vec512_ten256_h, vec512_ten256_l,
	vec512_ten384_h, vec512_ten384_m, vec512_ten384_l,
	vec512_ten512_3, vec512_ten512_2, vec512_ten512_1,
	vec512_ten512_0, vec512_ten1024_6, vec512_ten1024_5,
	vec512_ten1024_4, vec512_ten1024_3, vec512_ten1024_2,
	vec512_ten1024_1, vec512_ten1024_0, vec512_ten2048_13,
	vec512_ten2048_12, vec512_ten2048_11, vec512_ten2048_10,
	vec512_ten2048_9, vec512_ten2048_8, vec512_ten2048_7,
	vec512_ten2048_6, vec512_ten2048_5, vec512_ten2048_4,
	vec512_ten2048_3, vec512_ten2048_2, vec512_ten2048_1,
	vec512_ten2048_0): Add multiple quadword integer consts.

	* src/testsuite/vec_perf_f128.c: New file.
	* src/testsuite/vec_perf_f128.h: New File.

	* src/testsuite/vec_perf_f32.c (timed_setup_f32_transpose):
	New function.
